requirements.txt
streamlit
pillow
google-generativeai
langchain-google-genai
langchain-core
faiss-cpu
sentence-transformers
googletrans==4.0.2
numpy
